Add RELEASE_NATIVE_OMAPI build flag
This flag will determine whether the existing Java APK OMAPI service
is included in the build and used, or the Rust native OMAPI service is
included and used.  It must be a build flag for several reasons, the
most important that if native OMAPI is being used, it must be started
in the early_hal boot stage, before the dynamic flagging
infrastructure is available. Making an OMAPI service available to
early boot is the reason for using the new implementation.

Bug: 380331467
Test: N/A
Flag: RELEASE_NATIVE_OMAPI
Change-Id: I0531b7cb89e0112764a17af0d5192c935ad52dba