Revert trunk-staging sqlite to 3.44.3
Reverting trunk-staging sqlite to 3.44.3 in an attempt to see if that
fixes the crashes on x86_64 platforms.  (3.44.3 is the release on
trunk-food and next.)

This change updates the flag declaration to use the proper namespace
and the current default value.  It is hoped that these two changes
will make the flag ready for gantry.

Deleting the file from release/flag_values/trunk_staging should revert
the flag to the default value in the flag declaration.

Flag: EXEMPT bug-fix
Bug: 405376569
Test: presumbit
Change-Id: I8ca29cdbf6722a85c72cf4ec083bb345e4c3911d