Add build flag for storage files version.
v3 storage has redaction in the package table node. This is NOT the flag
that will be used to turn on/off redaction, just the flag that controls
the storage file version. Decoupling this from launching the change.
After this, to change the storage files, we don't need a new flag, we
will just launch the next value with this flag.

Test: n/a, adding build flag
Flag: Build.RELEASE_ACONFIG_STORAGE_VERSION
Change-Id: If3fb16a997b3f529043768c7c1b87414717110e9