Set base sdk version for bp2a
This was missed during finalization, setting it now for bp2a for
correctness. It defaults to the correct value on bp2a since if
RELEASE_PLATFORM_BASE_SDK_EXTENSION_VERSION is undefined then it uses RELEASE_PLATFORM_SDK_EXTENSION_VERSION instead.

It does matter for the sdk extension finalization though so this will
have to be cherry picked to the 25q2-sdk-dev

Test: m sdk sdk_repo dist
Bug: 409708158
Change-Id: I49ad01b228bf38cb6834a88e9b2edbb4cde6009c