Run connectivity tests in presubmit when ramping beta flags
There is nothing more frustrating than ramping a flag and getting the
flag ramp reverted 8 hours later because of a postsubmit failure. This
change configures to run the connectivity tests in presubmit when a
mainline beta flag is ramped to development.

That being said, I don't know how presubmit is configured and whether
these flags actually take effect during the presubmit run, so TBD on
whether this should be submitted.

Test: TH
Flag: NONE yes
Change-Id: I9749fdc4499c09bf7af95093b4f28c3e981b07db