Advance RELEASE_BUILD_CLANG_VERSION to r574158 for bp3a
git_main-throttled-monitored has a builder for bd6a but bd6a inherits
its compiler from bp3a which is r563880c. Warnings for r574158 will not
be understood by r563880c and cause build failures as seen in
b/456584904. We are updating bp3a to r574158 to avoid this issue but we
do not need to pull this commit to the release branch.

Bug: 456584904
Test: presubmit + ABTD Run
Change-Id: I0c3175d29cd6e9fedc5186c5ba85ce537fde01f1