Advance RELEASE_BUILD_CLANG_VERSION to r584948 for bp4a
The mainline release variant pulls directly from bp4a not cp2a
(as shown in mainline.textproto). So we will get build failures
similar to b/450669360 if we drop support of old clang in the
main branch. We can solve this by advancing bp4a to r58494 but
we do not want to pull this to the release branch.

Bug: 445718771
Test: presubmit
Change-Id: I0acabf71e06fe7a7f367112045c69cab689ad9f1